fix: correct func_flag type from u16 to u32 and remove broadcast
- Change func_flag type from u16 to u32 in FeiQPacket model
- Update parser to parse func_flag as u32
- Remove broadcast_online_presence function (moved to discovery module)
- Clean up unused broadcast code in main.rs

These changes align with the FeiQ protocol specification where
func_flag is a 32-bit value.

+++ src-tauri/src/network/feiq/model.rs
@@ -120,7 +120,7 @@ pub struct FeiQPacket {
     pub pkg_type: String,
     /// 功能标识位（如128=客户端在线广播）
     #[serde(default)]
-    pub func_flag: u16,
+    pub func_flag: u32,
     /// 原始MAC地址（如5C60BA7361C6）
     #[serde(default)]
     pub mac_addr_raw: String,
